clean Test
The BUILD-OPERATE-CHECK2 pattern is made obvious by the structure of
these tests. Each of the tests is clearly split into three parts. The
first part builds up the test data, the second part operates on that
test data, and the third part checks that the operation yielded the
expected results.