Bigger Selection
Even though this is close to a violation of the rule about mental
mapping, it seems appropriate in this case. Notice, once you know the
meaning, your eyes glide across that string and you can quickly
interpret the results. Reading the test becomes almost a pleasure.
Just take a look at Listing 9-5 and see how easy it is to understand
these tests.